feat: added docker and eslint configs
