+++ .dockerignore
@@ -1,0 +1,32 @@
+# docker related files
+.dockerignore
+Dockerfile
+
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+node_modules/
+.pnp/
+.pnp.js
+
+# production
+build/
+
+# misc
+CONTRIBUTING.md
+CONTRIBUTING.md
+LICENSE
+.eslintignore
+config.eslint.yml
+.DS_Store
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+.git/
+.gitignore
+README.md
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
