+++ docker-compose.yml
@@ -1,0 +1,7 @@
+version: "3.8"
+services:
+  web:
+    restart: on-failure
+    build: .
+    ports:
+      - "5000:5000"
