+++ package.json
@@ -1,27 +1,20 @@
 {
-  "name": "app1",
+  "name": "email-service-frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.2",
-    "@testing-library/user-event": "^12.3.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-scripts": "4.0.1",
-    "web-vitals": "^0.2.4"
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "4.0.0",
+    "serve": "^11.3.2"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "deploy:prod": "serve -s build",
     "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "lint": "eslint -c config.eslint.yml . --ext .jsx",
+    "lint:fix": "eslint -c config.eslint.yml . --ext .jsx --fix",
+    "start": "react-scripts start"
   },
   "browserslist": {
     "production": [
@@ -34,5 +27,14 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run lint"
+    }
+  },
+  "devDependencies": {
+    "eslint-config-airbnb": "^18.2.1",
+    "husky": "^4.3.0"
   }
 }
