+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "CSI-VIT",
+  "name": "Computer society of India, VIT student branch",
   "icons": [
     {
       "src": "favicon.ico",
