+++ src/pages/HomePage/HomePage.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Home() {
+    return (
+        <h1>Hello, world!</h1>
+    );
+}
+
+export default Home;
